all: clean up linter findings; add .golangci.yml
- .golangci.yml: skip generated internal/pb so the proto output stops
  drowning the report in noise.
- safefile.go: drop dead flag2str/prflag/xflag2str debug helpers;
  silence best-effort Close/Remove in cleanup().
- copy_mmap.go: deprecated os.SEEK_SET -> io.SeekStart.
- cmp/cmp.go: remove unused src/dst fields; collapse if-return.
- clone/hardlink.go: drop bogus "// go:build unix" (extra space made
  it a comment, not a directive; tag was redundant anyway).
- info_darbsd.go: zero-extend darwin int32 dev_t via uint32 step
  (no-op on freebsd) so the conversion intent is explicit.
- tests: drop three unused newBenchAsserter copies; check (or _-discard)
  Write/Sync/Unmap returns; fix ineffectual sf assignment.
- gosec #nosec annotations on the path-API Open/OpenFile call sites
  and the syscall.Mknod/dev_t conversions that gosec cannot prove safe.

+++ info_darbsd.go
@@ -27,8 +27,12 @@ func makeInfo(fi *Info, nm string, st *syscall.Stat_t, x Xattr) {
 		Ino:    st.Ino,
 		Siz:    st.Size,
 		Blocks: int64(st.Blocks),
-		Dev:    uint64(st.Dev),
-		Rdev:   uint64(st.Rdev),
+		// st.Dev/st.Rdev are int32 on darwin and uint64 on
+		// freebsd. Real device IDs are non-negative; the
+		// uint32 step makes the zero-extension explicit on
+		// darwin and is a no-op on freebsd.
+		Dev:  uint64(uint32(st.Dev)),  // #nosec G115 -- zero-extend darwin int32 dev_t
+		Rdev: uint64(uint32(st.Rdev)), // #nosec G115 -- zero-extend darwin int32 dev_t
 
 		Mod:   fs.FileMode(st.Mode & 0777),
 		Uid:   st.Uid,
